+++ src/client/components/cmSliding/cmSliding.js
@@ -14,7 +14,7 @@ Template.cmSliding.onCreated( function(){
     const self = this;
 
     // be verbose if asked for
-    if( cookieManager.conf.verbosity & CM_VERBOSE_COMPONENTS ){
+    if( cookieManager._conf.verbosity & CM_VERBOSE_COMPONENTS ){
         console.debug( 'pwix:cookie-manager cmSliding onCreated()' );
     }
 
@@ -43,7 +43,7 @@ Template.cmSliding.onRendered( function(){
     const self = this;
 
     // be verbose if asked for
-    if( cookieManager.conf.verbosity & CM_VERBOSE_COMPONENTS ){
+    if( cookieManager._conf.verbosity & CM_VERBOSE_COMPONENTS ){
         console.debug( 'pwix:cookie-manager cmSliding onRendered()' );
     }
 
@@ -52,7 +52,7 @@ Template.cmSliding.onRendered( function(){
 
     // display the adertising cookie message if we have no user consent, or too old
     const consent = cookieManager.consentRead();
-    const show = ( consent === null ) || ( Date.now() - consent.date > cookieManager.conf.consentLifetime );
+    const show = ( consent === null ) || ( Date.now() - consent.date > cookieManager._conf.consentLifetime );
     //console.debug( 'consent=', consent, 'show=', show );
     if( show ){
         self.$( '.cmSliding .cm-body' ).addClass( 'show' );
@@ -114,7 +114,7 @@ Template.cmSliding.events({
 
 Template.cmSliding.onDestroyed( function(){
     // be verbose if asked for
-    if( cookieManager.conf.verbosity & CM_VERBOSE_COMPONENTS ){
+    if( cookieManager._conf.verbosity & CM_VERBOSE_COMPONENTS ){
         console.debug( 'pwix:cookie-manager cmSliding onDestroyed()' );
     }
 });