+++ src/common/js/global.js
@@ -5,3 +5,7 @@
 cookieManager = {
     _conf: {}
 };
+
+// make the cookieManager available to each and every package
+console.log( 'attaching cookieManager to Meteor global...' );
+Meteor.cookieManager = cookieManager;