+++ package.js
@@ -1,6 +1,6 @@
 Package.describe({
     name: 'pwix:cookie-manager',
-    version: '1.2.0-rc',
+    version: '1.2.0',
     summary: 'Cookies management for Meteor',
     git: 'https://github.com/trychlos/pwix-cookie-manager',
     documentation: 'README.md'
@@ -40,6 +40,7 @@ function configure( api ){
     api.use( 'pwix:i18n@1.0.0' );
     api.use( 'pwix:layout@1.2.0' );
     api.use( 'pwix:modal@1.4.0' );
+    api.use( 'pwix:toggle-switch@0.3.0' );
     api.use( 'tmeasday:check-npm-versions@1.0.2', 'server' );
     api.addFiles( 'src/client/components/cmSliding/cmSliding.js', 'client' );
     api.addAssets([