+++ ChangeLog.md
@@ -17,6 +17,7 @@
     - Replace merge depency with lodash
     - Rename conf to _conf making clearer this is a private variable (todo #21)
     - Merge config.js + defaults.js into configure.js (todo #22)
+    - configure() now acts both as a getter and a setter (todo #23)
 
 ### 1.1.1